# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  /** site_visit | consultation | planned_work | service_call | change | other */
  type: string;
  /** @nullable */
  clientSite?: string | null;
  /** ISO date (YYYY-MM-DD) */
  date: string;
  /**
     * HH:MM
     * @nullable
     */
  startTime?: string | null;
  /**
     * HH:MM
     * @nullable
     */
  endTime?: string | null;
  /** planned | in_progress | done | cancelled */
  status: string;
  /** @nullable */
  assignedPersonId?: number | null;
  /** @nullable */
  assignedPersonName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  transportKm?: number | null;
  /** @nullable */
  transportCost?: number | null;
  /** @nullable */
  fines?: number | null;
  /** @nullable */
  parking?: number | null;
  taskCount?: number;
  taskDoneCount?: number;
  attachmentCount?: number;
  createdAt: string;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  type: string;
  /** @nullable */
  clientSite?: string | null;
  date: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  status: string;
  /** @nullable */
  assignedPersonId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  transportKm?: number | null;
  /** @nullable */
  transportCost?: number | null;
  /** @nullable */
  fines?: number | null;
  /** @nullable */
  parking?: number | null;
}

export interface JobUpdate {
  /** @minLength 1 */
  title?: string;
  type?: string;
  /** @nullable */
  clientSite?: string | null;
  date?: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  status?: string;
  /** @nullable */
  assignedPersonId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  transportKm?: number | null;
  /** @nullable */
  transportCost?: number | null;
  /** @nullable */
  fines?: number | null;
  /** @nullable */
  parking?: number | null;
}

export interface JobStatusUpdate {
  status: string;
}

export interface Task {
  id: number;
  jobId: number;
  title: string;
  done: boolean;
  /** Marks this task as a change request / extra work */
  isChangeRequest: boolean;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  isChangeRequest?: boolean;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  done?: boolean;
  isChangeRequest?: boolean;
}

export interface Attachment {
  id: number;
  jobId: number;
  /** photo | delivery_note | receipt | invoice | manual_item */
  type: string;
  /** @nullable */
  fileName?: string | null;
  /**
     * base64 data URL or file path
     * @nullable
     */
  url?: string | null;
  /** @nullable */
  description?: string | null;
  /**
     * For manual items and invoices
     * @nullable
     */
  amount?: number | null;
  createdAt: string;
}

export interface AttachmentInput {
  type: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
}

export interface Person {
  id: number;
  name: string;
  createdAt: string;
}

export interface PersonInput {
  /** @minLength 1 */
  name: string;
}

export interface DashboardSummary {
  todayCount: number;
  weekCount: number;
  plannedCount: number;
  inProgressCount: number;
  doneCount: number;
  totalHoursThisWeek: number;
  totalRevenueThisWeek: number;
}

export type ListJobsParams = {
/**
 * ISO date string (YYYY-MM-DD)
 */
from?: string;
/**
 * ISO date string (YYYY-MM-DD)
 */
to?: string;
status?: string;
/**
 * @nullable
 */
assignedPersonId?: number | null;
};