# lib/api-zod/src/generated/types/attachmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AttachmentInput {
  type: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
}