# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todayCount: number;
  weekCount: number;
  plannedCount: number;
  inProgressCount: number;
  doneCount: number;
  totalHoursThisWeek: number;
  totalRevenueThisWeek: number;
}