# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  title: string;
  /** site_visit | consultation | planned_work | service_call | change | other */
  type: string;
  /** @nullable */
  clientSite?: string | null;
  /** ISO date (YYYY-MM-DD) */
  date: string;
  /**
     * HH:MM
     * @nullable
     */
  startTime?: string | null;
  /**
     * HH:MM
     * @nullable
     */
  endTime?: string | null;
  /** planned | in_progress | done | cancelled */
  status: string;
  /** @nullable */
  assignedPersonId?: number | null;
  /** @nullable */
  assignedPersonName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  transportKm?: number | null;
  /** @nullable */
  transportCost?: number | null;
  /** @nullable */
  fines?: number | null;
  /** @nullable */
  parking?: number | null;
  taskCount?: number;
  taskDoneCount?: number;
  attachmentCount?: number;
  createdAt: string;
}