# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  /** @minLength 1 */
  title: string;
  type: string;
  /** @nullable */
  clientSite?: string | null;
  date: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  status: string;
  /** @nullable */
  assignedPersonId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  transportKm?: number | null;
  /** @nullable */
  transportCost?: number | null;
  /** @nullable */
  fines?: number | null;
  /** @nullable */
  parking?: number | null;
}