# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stavba Job Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Person {
  id: number;
  name: string;
  createdAt: string;
}